Move model-specific utils to `feature/models`
